+++ packages/core/src/message/chat.ts
@@ -1,0 +1,8 @@
+import { ChatMessage } from "@windflow/core/types"
+export class ChatMessageManager {
+  constructor() {}
+  addMessage(message: ChatMessage) {
+    // create index
+    //
+  }
+}