+++ packages/core/src/message/recorder.ts
@@ -1,9 +1,0 @@
-export interface RecorderRules{
-
-}
-export class Recorder<K, V> {
-  #record: Map<K, V>
-  constructor() {
-    this.#record = new Map()
-  }
-}